+++ src/artifact/artifact-info.cpp
@@ -5,8 +5,6 @@
 
 #include "artifact/artifact-info.h"
 #include "artifact/random-art-effects.h"
-#include "cmd-item/cmd-smith.h"
-#include "mind/mind-weaponsmith.h"
 #include "object-enchant/activation-info-table.h"
 #include "object-enchant/object-ego.h"
 #include "object-enchant/object-smith.h"
