--- src/cmd-item/cmd-smith.h
@@ -1,10 +1,0 @@
-﻿#pragma once
-
-#include "system/angband.h"
-
-// @todo 過去のセーブファイルからの移行用コードのコンパイルのためとりあえず残す
-// 最終的には鍛冶師ではリファクタリングの前後でのセーブファイル移行を不可にする
-#define MIN_SPECIAL_ESSENCE 200
-
-struct player_type;
-void do_cmd_kaji(player_type *creature_ptr, bool only_browse);
