+++ src/flavor/object-flavor.cpp
@@ -10,7 +10,6 @@
  */
 
 #include "flavor/object-flavor.h"
-#include "cmd-item/cmd-smith.h"
 #include "combat/shoot.h"
 #include "flavor/flag-inscriptions-table.h"
 #include "flavor/flavor-util.h"
