+++ src/io/input-key-processor.cpp
@@ -39,7 +39,6 @@
 #include "cmd-item/cmd-quaff.h"
 #include "cmd-item/cmd-read.h"
 #include "cmd-item/cmd-refill.h"
-#include "cmd-item/cmd-smith.h"
 #include "cmd-item/cmd-throw.h"
 #include "cmd-item/cmd-usestaff.h"
 #include "cmd-item/cmd-zaprod.h"
@@ -74,6 +73,7 @@
 #include "mind/mind-elementalist.h"
 #include "mind/mind-magic-eater.h"
 #include "mind/mind-sniper.h"
+#include "mind/mind-weaponsmith.h"
 #include "mind/snipe-types.h"
 #include "player-status/player-energy.h"
 #include "player/attack-defense-types.h"
