+++ src/store/store-key-processor.cpp
@@ -14,7 +14,6 @@
 #include "cmd-io/cmd-lore.h"
 #include "cmd-io/cmd-macro.h"
 #include "cmd-item/cmd-equipment.h"
-#include "cmd-item/cmd-smith.h"
 #include "cmd-visual/cmd-draw.h"
 #include "cmd-visual/cmd-visuals.h"
 #include "game-option/birth-options.h"
@@ -23,6 +22,7 @@
 #include "io/input-key-requester.h"
 #include "mind/mind-elementalist.h"
 #include "mind/mind-sniper.h"
+#include "mind/mind-weaponsmith.h"
 #include "store/home.h"
 #include "store/museum.h"
 #include "store/purchase-order.h"
